% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresstest.R
\name{apply.argset}
\alias{apply.argset}
\title{tests a function provided to the argument FUN with argument value sets
prepared earlier and stored in the default work environment and returns 
a vector with {PASS|FAIL} for every argument value combination.}
\usage{
apply.argset(env = NULL, arg_register = cont.env$arg_register, FUN,
  subset = NULL, verbose = FALSE, DEBUG = FALSE)
}
\arguments{
\item{env}{work environment (if NULL, uses the default)}

\item{arg_register}{an argument register that contains all argument 'states'}

\item{FUN}{a name of the function to be tested (only as a character string)}

\item{subset}{a vector with numbers of argsets to be applied to the function
(i.e. only combinations with those numbers will be tested)}

\item{verbose}{provides additional text output during processing}

\item{DEBUG}{enters the debug mode on function entry}
}
\description{
An alternative work environment may be supplied in the future 
(not tested yet)
}
\author{
cloudcell
}

