% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzdemofunc.R
\name{fuzzdemofunc}
\alias{fuzzdemofunc}
\title{Generates errors for several combinations of input parameters to test the
existing and emerging functionality of the package}
\usage{
fuzzdemofunc(x, y, option, suboption)
}
\arguments{
\item{option}{any character value from "a", "b", "c"}

\item{suboption}{any character value from "a", "b", "c", "d"}

\item{x:}{any numeric scalar value (non-vector)}

\item{y:}{any numeric scalar value (non-vector)}
}
\description{
Whenever options lead the control flow within a function to a 'demo bug', 
the function stops and the test framework records a 'FAIL' result.
Upon a successful completion, the function returns a numeric value into the 
environment from which the function was called.
}
\author{
cloudcello
}

