% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stresstest.R
\name{generate.argset}
\alias{generate.argset}
\title{Generates a set of all the combinations of arguments (argset) to be used
to test a selected function}
\usage{
generate.argset(arg_register, cust.env = NULL, verbose = FALSE,
  DEBUG = FALSE, display_progress = FALSE)
}
\arguments{
\item{arg_register}{an argument register that contains all argument 
'states' required to be tested (while combined with 
'states' of other arguments)}

\item{cust.env}{custom work environment}

\item{verbose}{provides additional text output during processing}

\item{DEBUG}{enters the debug mode on function entry}

\item{display_progress}{prints "." on each iteration in the standard output}
}
\description{
This function uses an argument 'register' created earlier and generates all
possible combiations of arguments, including argument 'states', such as
'missing' specified as '__MISSING__' in the register. The result is stored
in the work environment. There is no special function to create an
argument register to be used with 'generate.argset()' as the register is
simply a list. Please see file 'fuzzdemo.R' for an example. 
At the moment the following character strings are interchangeable when
setting up the test (the argument register):
"__val.NULL__" and "NULL", 
"__val.NA__" and "NA",
"__val.TRUE__" and "TRUE",
"__val.FALSE__" and "FALSE".
If you would like to actually assign character strings such as "NULL", "NA",
"TRUE", "FALSE", simply override the function "prepareArgs" with your
modified version.
}
\author{
cloudcell
}

