% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lmessage}
\alias{lmessage}
\title{displays a message and stores it in a log file}
\usage{
lmessage(x, .logger_name = "fuzzlog", .env = cont.env, .verbose = TRUE)
}
\arguments{
\item{.env}{container-environment of the testing framework}

\item{.verbose}{(default==TRUE) whether to call message}

\item{...}{values to be concatenated and submitted as text to message()}

\item{.logger}{logger name/handle (as a character string), which must 
be set up prior to running lmessage()}
}
\description{
No matching of partially spelled arguments is performed in this function.
}

