% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logger_new}
\alias{logger_new}
\title{creates a new logger}
\usage{
logger_new(lgr_name, dest = getwd(), fname = NULL, env = cont.env)
}
\arguments{
\item{lgr_name}{logger name}

\item{dest}{destination directory (defaults to the current work directory)}

\item{env}{work (container) environment (default 'cont.env')}
}
\description{
creates a new logger
}

