% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ls_loggers}
\alias{ls_loggers}
\title{lists existing loggers}
\usage{
ls_loggers(env = cont.env)
}
\arguments{
\item{env}{work (container) environment (default == 'cont.env')}
}
\description{
lists existing loggers
}

