% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_logger}
\alias{new_logger}
\title{creates a new logger}
\usage{
new_logger(fname = NULL, logger_name = "fuzzlog", dest = getwd(),
  env = cont.env)
}
\arguments{
\item{fname}{log file name}

\item{logger_name}{logger name (character identifier / handle)}

\item{dest}{destination directory (defaults to the current work directory)}

\item{env}{work (container) environment (default 'cont.env')}
}
\description{
creates a new logger
}

