% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_tests}
\alias{plot_tests}
\title{Plots test results}
\usage{
plot_tests(env = cont.env, pass = TRUE, fail = TRUE, dist = 1,
  verbose = FALSE, DEBUG = FALSE)
}
\arguments{
\item{env}{work environment, set only if default environment is not used}

\item{pass}{determines whether the chart includes the full test set data.
I.e. whether all of the 'PASS' tests are included (default is 
TRUE). If set to FALSE, all the PASS data will not be drawn, 
which sometimes might speed up the drawing process.)}

\item{fail}{determines whether the chart includes the full test set data.
I.e. whether all of the 'PASS' tests are included (default is 
TRUE). If set to FALSE, all the PASS data will not be drawn, 
which sometimes might speed up the drawing process.)}

\item{dist}{the minimum distance between two adjacent lines from different
'levels' within the same 'category' at a point of intersection 
with the category ('argument') axis. The default of 1.0
makes a gap between any two adjacent groups of lines 
sufficient for a maximum possible number of lines to pass through 
without overlapping adjacent lines at the point of intersection
with the axis. Distance of 0 makes no gap (with only a 
regular distance separating any two adjacent lines at a point
of intersection with the vertical axis). A dist of -0.9999 
creates almost complete overlap with an adjacent group.
The value of dist must be greater than -1. 
(Ror each particular axis, the distance between groups of lines
 representing test cases is based on a product of the number 
 of groups of all the neighboring axes.)}

\item{verbose}{provides additional text output during processing}

\item{DEBUG}{enters the debug mode on function entry}
}
\description{
uses parallelplot with added spacing to diplay combinations of argument 
values groupped by PASS/FAIL exit state
}
\author{
cloudcell
}

