% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_logger}
\alias{rm_logger}
\title{removes a specified logger after flushing and closing its log file}
\usage{
rm_logger(logger_name = "fuzzlog", env = cont.env)
}
\arguments{
\item{logger_name}{logger name}

\item{env}{work (container) environment (default == 'cont.env')}
}
\description{
removes a specified logger after flushing and closing its log file
}

