% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsummary.R
\name{summary_fuzztest}
\alias{summary.fuzztest}
\alias{summary_fuzztest}
\alias{test_summary,}
\title{Aggregates all the test results by argument 'state' and provides statistics}
\usage{
summary_fuzztest(env = cont.env, verbose = FALSE, DEBUG = FALSE)
}
\arguments{
\item{env}{work environment, set only if default environment is not used}

\item{verbose}{provides additional text output during processing}

\item{DEBUG}{enters the debug mode on function entry}
}
\description{
The Function splits statistics into detailed and short summaries.
The 'detailed statistics': PASS/FAIL count, percentage of failed tests.
The short summary contains only key data for determining a degree to which 
some argument 'state' is associated with whether a function completes
its operations with a 'PASS' or 'FAIL' state. Use this statistics to isolate
those arguments analyze how code 'misbehaves' by looking up a more detailed
output from the first 'detailed statistics' table.
}
\author{
cloudcell
}

