% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lmessage}
\alias{lmessage}
\title{displays a message and stores it in a log file}
\usage{
lmessage(x, logger, env = cont.env, verbose = TRUE)
}
\arguments{
\item{x}{value to be submitted as a message to message()}

\item{logger}{logger handle which must be set up prior to running 
lmessage}

\item{env}{container-environment of the testing framework}

\item{verbose}{(default==TRUE) whether to call message}
}
\description{
displays a message and stores it in a log file
}

