% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_logger}
\alias{rm_logger}
\title{removes a specified logger after closing the log file}
\usage{
rm_logger(lgr_name, env = cont.env)
}
\arguments{
\item{lgr_name}{logger name}

\item{env}{work (container) environment (default 'cont.env')}
}
\description{
removes a specified logger after closing the log file
}

